requirements.txt
##python3.6
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
clang==5.0
cmake==3.26.3
colorama==0.4.5
contextlib2==21.6.0
cycler==0.11.0
Cython==0.29.35
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dlib @ file:///D:/bld/dlib_1617441818780/work
docutils==0.18.1
entrypoints==0.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.4
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==23.4.1
kiwisolver==1.3.1
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy @ file:///D:/bld/numpy_1626682185159/work
oauthlib==3.2.2
onnxruntime==1.10.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.9.6
prometheus-client==0.17.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==1.1.6
pyzmq==25.1.0
qtconsole==5.2.2
QtPy==2.0.1
readme-renderer==34.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
scipy==1.5.4
Send2Trash==1.8.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tf-slim==1.1.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
twine==3.8.0
typing-extensions==3.7.4.3
urllib3==1.26.16
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.4
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
